backend/requirements.txt
pandas
numpy
vaderSentiment
textblob
transformers
praw
requests
spacy
scikit-learn
scipy
google-generativeai
geopy
pycountry
google-play-scraper
sys